chore(evm): move wrappers utils to types pkg (#2905)
* chore(evm): move wrappers utils to types pkg

* run make format

---------
